fix: Do not display the delete link button and cross for reverse links
This closes request #29260 Reverse links must not be selected for deletion

Given an artifact with artifact links and some reverse links, if we display the reverse links by type,
then the cross to delete this link is display.
It does nothing because deleting a reverse link is not possible.
This action is not displayed for reverse links displayed by tracker.

To be consistent, we do not have to display this cross.

Change-Id: I70b41d431ee5163dba03013781b6818e8c964474