Extract the download URL from ONLYOFFICE callback save request
The goal of this contribution is to put in place the infrastructure
required to parse the ONLYOFFICE callback requests. For now we only log
this information. Additional contributions will download the document
and save it and extract more information from the callback requests.

To test:
1. Start ONLYOFFICE in your dev stack
2. Enable the feature flag with `tuleap config-set feature_flag_onlyoffice_edit_document 1`
3. Open a document you can edit into ONLYOFFICE
4. Do some changes in the document and then close ONLYOFFICE
5. After a few seconds you should see the download URL of the edited
   document in the logs

Part of story #28263: edit office document stored in Document

Change-Id: I75ea62019059c1fdb8d9ba1d7089ad5424121359