fix: Disabled drag and drop of image when editing an existing followup comment
This closes request #29267 Disabled drag and drop of image when editing an existing followup comment

When editing a followup comment, it's not possible to copy&past an image but we can drag&drop it.
When saving the new version of the followup, the image is not saved.

This is because of the backend that handles the images in HTML content.
We have to save images in the Attachment field.
However, when editing, this means to rewrite some history to have the new images.
This is too dangerous.

We have to disable copy&past and drag&drop during the followup edition.

How to test
----------

* Switch an existing followup comment in edition
* Switch to HTML format if needed
* Try to drag&drop an image

=> An error message saying we are not allowed to do that must be
displayed

Change-Id: If618bd1b7b259ec346b93ff88d4472caced9686e