chore: Git permissions-per-group
How to test:
- In a project with Git service active, go to Project Administration >
Permissions > Per Group
- When you click the "See all permissions" button, the Vue app is loaded
and works as before.

Notes:
I tried using Vite instead of webpack, but a piece is missing. We might
need a plugin to hammer the gettext po file to the right format. Out of
the box, the PO file is loaded but translations are still in english.
I'm guessing some issue is silenced. Anyway, Vite usage can come later.

part of request #29265 Split the build of git plugin frontend assets
into multiple apps

Change-Id: Ia928810fc8620bc877f23f8b86fc13dfaf9f069c