Closes request #29269: Script collecting JS errors should be loaded first
Note that the script at a later stage on BurningParrot pages than on
FlamingParrot pages. BP pages does not offer the possibility to load
scripts in the <head> so even if they have script tags in their bodies
for some (bad) reasons the code is more likely to be forced to wait than
everything is loaded before doing something.

Change-Id: Ibd7c094e4b21127fe453e3f6e8d4b313253e0d13