fix: document cascade metadata update
This closes request #29275 Recursion option is not taken into account at folder metadata edition

When editing a folder metadata and applying a recursion option to cascade these changes,
this recursion option is not taken into account and only the current folder metadata are edited

How to test
-----------

* Disable in your Document administration the status metadata
* Update a folder metadata and apply a recursion option (by selecting
  some and unselecting the others)

=> Metadata must be updated according to the recursion option

* Enable in your Document administration the status metadata
* Update a folder metadata and apply a recursion option (by selecting
  some and unselecting the others)

=> Metadata must be updated according to the recursion option

Note: Updating Metadata without status resets the status to "None".
This is the current behavior. It will be fixed in a dedicated request if
needed.

Change-Id: I38e0ef04b0c339969724635a3bc5689efe89a0fa