feat: Initiate display of Source column
User should be able to identify File versions that have been updated by
ONLYOFFICE integration.

This contribution display such column but with common value "Uploaded"
since save from ONLYOFFICE is not implemented yet. If
tuleap-plugin-onlyoffice is not enabled, then the column is not
displayed.

Part of story #28263: edit office document stored in Document

Change-Id: If479e69a0494ed07aba0edd3cfb9c4626e87400b