Close request #29278: Let the FTS backends deal with the transformation of the usual content format
This moves the conversion of usual content formats (HTML, CommonMark and
plaintext) to the FTS backends themselves. This present multiple
advantages:
* it avoids CPu cycles if there is no FTS backends actived or usable
* it moves the conversion into the async code which can improve the
  experience for the user when updating items
* it lays the groundwork for expanding the usage of the FTS outside of
  the tracker plugin, having a common way to manage the usual content
  will reduce the code duplication between the different plugins
* it also does not assume the capabilities of the FTS backends, some
  might have the capability deal with it directly (e.g. Elastic can
  strip HTML tags when indexing documents)

No functional changes expected.

Note that the step definition fields are still pre-processed into plaintext
due to their unusual format. They are unique in the sense that for a
same item to index we might have multiple formats.

Change-Id: I493ed0337d76d75c422eb6310d323ec7cf10b23b