Drop strict restriction on the expected version that should be updated
Due to the async nature of the code we cannot really give feedback in a
way that would allow the user to remediate the situation.
Worst case scenario if another document has been uploaded in the
meantime it will still be accessible via the logs/history.

Part of story #28263: edit office document stored in Document

Change-Id: I703ea313f7352910e6c87511c119afad713d22ab