fix: Unused status metadata must not be cleared at metadata edition
This is part of request #29282 Item metadatas cleared at update when metadata is not used

Given a document item (either a folder or any other file type) that has a status defined with the status metadata,
if a project administrator disables the usage of the status metadata
and then any other metadata for this item is updated, then the status is reset to none, which is wrong.

The status value must not be cleared when ``status``` metadata is no more used.
(this worked like this in the legacy docman)

How to test
----------

* Go to a document service that has the status metadata enabled
* Edit the metadata of any item and set a status
* Disable the usage of the status metadata in the document
  administration
* Edit any other metadata of the previous item

=> The status metadata value must not be reset to none
(this can be checked by either querying the database or enabling the
status metadata and opening the edit metadata modal of the item).

Change-Id: I4b216a540cefcfcfec610322faf0d75ad66fe801