Use OpenSSL 3.0.7 to build the `tuleap-git-bin` package
Another solution would be to update our global pin but this will wait
we cut the 14.2 release so we can have some time to test everything is
fine.

To test: rebuild and deploy the package on your dev instance. No changes
are expected (especially given the fact no operation are made against
HTTP remotes).

Closes request #29271: Remediation for CVE-2022-3786 & CVE-2022-3602 (OpenSSL 3.0.0 to 3.0.6)

Change-Id: Ic58bc24f0e8cf4c891a066613244aad187b4d7ff