feat: Display [Reply] button only on latest comment
part of story #29240 reply to pull request comment

We now only display the [Reply] button on the latest comment of a
thread. The idea is to prevent users to think they can answer to a particular comment
in a thread to create a sub-discussion. Since there is nothing else in
the footer but the reply button, we don't display it at all.

How to test:
- Create a new comment
--> It has a footer containing the [Reply] button

- Reply to this comment
--> The root comment has no footer anymore
--> The new comment has a footer containing the [Reply] button

- Reply once again to the comment
--> The root comment has no footer
--> The previous comment has no footer anymore
--> The new comment has a footer containing the [Reply] button

Change-Id: Iec697b499bf2bd210d028ef8151fcb247d69a650