File documents are displayed at first when user sorts the Filename column
Part of request #29209 Sort on Created Date and Filename doesn't work

At the moment, when clicking on the the Filename column the item
returned are sorted by item type and filename, in that case, the search
report displayed might not be totally clear since you can have all
others item type displayed before the files.
This contribution aim to put all the file documents always at first places when the
user want to sort the Filename column.

Change-Id: I3abfa7d7407362c6efbca5fef4d0cc0c48230e2c