fix: Tooltips are not loaded properly
part of story #29240 reply to pull request comment

Tooltips were loaded by angular using $timeout. It worked when the
comments were displayed by some angular components, but since we have
now an Hybrids custom element, it is working 1 time out of 2.

The solution brought by this contribution is to let the Hybrids
component load the tooltips when it did its first rendering.

Please note that if you add a reply containing a reference, it won't
setup a tooltip for it. The reason is that we cannot call loadTooltips
several times, or we end up having X tooltips stacked up when we hover
the reference. Please also note it does work for new global and inline
comments, as once they are created, the tooltips are created right away.

Change-Id: I213a8466d6e575577f5bd691d5e38102286dd23e