Implement request #29601 Use label for GitLab group selection
How to test:
With Git and GitLab plugin, go to a project using the Git service, then
go to Git Administration > GitLab group link tab
Fill the first step of the wizard (gitlab URL and token).
The app will fetch the list of GitLab groups you can see with the token.
If you click on the group name, it will select the matching radio button
on the same line.

Change-Id: I567f0aae07c96c6b2fb33ca3065255a3be467b3a