refactor: Migrate side-by-side-line-grouper to TypeScript
part of request #29602 Migrate pullrequest file-diffs to TypeScript

[Bonus]

The algorithm has been simplified. Before, we were mutating the file
lines by adding them information regarding the group they belong. This
property is only used to build the groups of lines, and it was pretty
simple to get rid of it by fetching the last group constructed during
the reduce function.

-- No functional change expected

Change-Id: Ibcef465ac0a94a11847e30ab3c3477e2947fc0a6