feat: account creation confirmation to BP
When a user register a new account and there is no need for site admin
approval, then the confirmation page is now displayed with BurningParrot.

This contribution only targets the step "O" in the flow chart below.

If an error during the delivery of the confirmation mail occurs then an
error feedback is displayed but the content of the page stay the same.
It will be improved later to not display error "The mail was not
accepted for the delivery…" next to "Confirmation link sent".

```mermaid
graph TD
    A[Register] --> B{Captcha <br>ok?}
    B -->|no| F
    B -->|yes| D(Create account)
    D --> E{Account <br>created?}
    E -->|no| F[Errors!]
    E -->|yes| G(OIDC link registering account)
    F --> H[fa:fa-star Display register form]
    H --> A
    G --> I{User created <br> by site admin?}
    I -->|yes| J{Should send <br>an email?}
    J -->|no| R
    J -->|yes| K(Send email to user with login and pwd fa:fa-bomb)
    I -->|no| L{sys_user_approval = 0?}
    L -->|no| M[fa:fa-star Your account has been created<br>it must be approved by an administrator<br>Once approved clic the confirm button in this email]
    L -->|yes| Q(Send confirmation email fa:fa-bomb)
    Q --> O[fa:fa-star Your account is created, </br>it now needs to be activated.<br>To do this clic the confirm button in the email]
    K --> R[fa:fa-star The user account was created.<br>Please note the login and pwd]

    style H fill:#ffe5e5,stroke:#f02727,color:#b70d0d
    style R fill:#ffe5e5,stroke:#f02727,color:#b70d0d
    style M fill:#ffe5e5,stroke:#f02727,color:#b70d0d
    style O fill:#f4f8e9,stroke:#6abf1d,color:#137900
```

Implementation notes:

For now code is kept in register.php but once other pages will be
converted to BurningParrot, they will be moved in a dedicated
DispatchableWithBurningParrot controller.

Change-Id: Ib72e9d847f63d2d5c6548d721f9d47a168638b4e