refactor: Migrate side-by-side-line-mapper to TypeScript
part of request #29602 Migrate pullrequest file-diffs to TypeScript

In this refactoring, I also introduce a file-line-helper which provides
type guards function to tell whether a line is unmoved, added or
removed, so we can make sure the line we are processing has the right
type while DRYing some code.

note: We may need a stub for LineHandles, but it is too early to
introduce it in unit tests. For now I'l keep things as they are.

-- No functionnal change expected

Change-Id: I733ce5a1d1a3f4f490acc2803f7d28fd27a2c77f