fix: GMT is not listed in available user timezones
This closes request #29606 GMT is not listed in available user timezones

Given a user with the GMT timezone, when this user accesses to its personal settings,
then another timezone is automatically displayed in the list, which leads to confusion.

No value must be displayed if the timezone is not listed in the user preferences.

How to test
-----------

* Login as a user which uses the GMT timezone
* Go to your personal settings

=> No timezeone must be selected and a warning must be displayed.

Technical note: Default timezone for new users from LDAP is now 'UTC'.
For new Tuleap installation, the default timezone for None user and
admin user is now 'UTC'.

Change-Id: Iacc8e56ef8099cd1d2f7979e81521affdc4eea7a