chore: Let Git commands failures exceptions be logged
This is part of request #29608 Let Git commands failures exceptions be logged

Given a Git command that ends in a non expected state (for example a failure in the command),
then this error will generate an Exception. This exception must not be caught.
It must ends with an error 500 and the exception will be logged in the
rest_api_log and/or in the nginx/fpm error log.

Git command failure must not occur.
If it happen, we need to have details about the command and the failure.

Change-Id: Ib10b8afdf29f05af1138503ff5a1edcfc37f85d3