chore: follow-up comments in hybrids
Comments order should now match the one in Tracker artifact view
(previously it was reversed).
There is now a little spinner icon to show that the comments are
loading.
The error message when loading comments shows up in the modal feedback
section (which is visible even when scrolling to the bottom). It is also
a bit more detailed.

No other functional change expected in edition modal:
- When there are no comments, an empty state is shown.
- Relative date / absolute date user preference is respected.
- Edited comments show the user editing them and the last update date.
- Anonymous users don't have a comment input (they cannot write a new
  comment).
- Tooltips are still shown on existing comments only (not new comments).
  They are not more broken than before.
- New comments are still saved.
- It's still possible to upload an image in HTML comment (with a File
  field present in the tracker).

part of request #22648 Convert AngularJS directives and controllers to
hybrids web components

Change-Id: I1efcef35b5b17272f4cf9cfd68b8212252642578