feat: Simplify new item creation modal
In the context of ONLYOFFICE integration, the creation of spreadsheet,
word, or presentation documents forces us to update the item creation
modal: include of OO types in the existing type selector at the top of
the modal would be complicated.

Therefore we now remove this type selector and the user must select the
right type of item beforehand, via a dropdown.

This dropdown is dispayled as a regular dropdown under a dedicated "+
New" button (folder header, quick look, empty state), or as a
sub-dropdown when the "+ New" is in the existing dropdown (e.g tree
view, search).

In order to not clutter this review, please note that:
* Minimal effort has been made in the styling (ellipsis button, caret
for submenu, submenu position, …). Application should not appear broken
but needs polishing.
* QuickLook action buttons need homogenization between item types
* Creation of a new version for an empty document should be based also
of the new dropdown

Part of story #28273: create OnlyOffice documents

Change-Id: I695352bd30bddda5f9388c86628a8b2fb28633e5