refactor: Hybrids.js placeholder + stronger widgets types
part of request #29602 Migrate pullrequest file-diffs to TypeScript

In this contribution, placeholders are now hybrids custom elements.
They can be either a code placeholder or a comment placeholder.
This improvement had to be made in order to strenghten widgets types.

Now we have three types of widgets:
- inline-comment widgets
- new-inline-comment-widgets
- placeholder-widgets

Since we had to check in several files wether a given widget is a
comment placeholder or a comment widget, these types come in handy and
make the code a bit simpler (production and test code).

How to test:
- No functional change expected, you should have a file diff with properly
sized code and comments placeholders, even if you new comments or
replies to the file.

Change-Id: I02ab9202625da8537b4ed3394f42acd3f4b0fbec