refactor: Use FileLine and GroupOfLines stubs in unit tests
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This refactoring is needed to ease the writing of unit tests in the
file diff feature. Moreover, it will be less work to do when we'll have
to update the shape of these objects, since there stubs are build at one
place.

It also makes the tests easier to understand. Indeed, we have test cases
with unmoved, added and removed lines and groups of lines, and it is
confusing to have only offsets (new/old/unidiff) to tell them appart.

-- No functional changes expected
-- CI should be happy

Change-Id: I76f692a8c7dc75c9f591fa79ca5ebd11c3f58b5b