refactor: Extract codemirrors init from lines-state
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This code inits the content of the two codemirrors and it is not the
state's role to do it. Moreover, it forced it to have a dependency
on codemirror Editor, and it was not great.

Now we have a code-mirror-content-manager doing it for us.

The two methods getRightLine and getLeftLines have been moved to this
object since they were only used in the side-by-side-diff-component.

-- no functional change expected

Change-Id: I3bf181d22d391f4249c2e84c6c38aff6f7941294