refactor: Migrate the file lines state + placeholders builders to TypeScript
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This one was hard to migrate, because the state was composed of a
multitude of exported function looking for stuff in some maps that was
initialized at some point. The placeholders builders had to be migrated
in the same patch, since the state is now an object providing those
functions they need to work.

Moreover, the way the state was mocked in the unit tests was not good.
Indeed, I have found a test that is influencing artificially its
result to make it pass. I don't have a solution yet to fix it. I've
spent 3 hours trying to debug it, but given I had no clue of what's the
actual expected behavior, I couldn't get any further.

The unit tests can now use a FileLinesStateStub to test their code if
they need. This stub has been designed to work with the real state
object, but while providing stubbed lines/handles/groups of lines.
This stub is able to prepend missing lines to the collection of file
lines, in order to be able to test a case where a group of lines is in
the middle of the file. Firsts and last lines of diffs being subject to
bugs, It allows us to replicate a given file-diff state with enough fidelity.

-- No functional change expected

Change-Id: Ieb67e7b75b3b28630e0c600f57deb40213b2586e