Fix request #28299 Absolute links opens a new tab
In order to test you can either rebuild the mediawiki package or
edit /usr/share/mediawiki-tuleap-flavor/LocalSettings.Tuleap.php and do
manually the change in this commit.

Please note that you might also need to edit the page you make your
tests on because it seems there is a cache on the generated HTML and
without edit, the link will still have a `_blank` attribute.

Change-Id: I3dc6b6ab9456637d153cc41a66361261cd711463