feat: Add a feature flag for history and search in artifact links
How to test:
- without changing the feature flag, there is no change in behaviour

$ tuleap config-set feature_flag_history_and_search_in_modal 1

- after enabling the feature flag with the command above, in Kanban,
Planning view and Test Management, when you open the artifact modal (to
create or edit an artifact), you should see a console info message
"Search is enabled".
The message will be removed on a later contribution, its only purpose is
to check that the feature flag is correctly transmitted.

part of story #24971 [modal] search for artifacts

Change-Id: I5bf1773b5bab8bc0945e34afae7b321492f94008