feat: Add type and per_type_id properties to History entries
- call GET /users/:id/history
- Entries have two more properties: type which represents the type of
item ("kanban" | "artifact") and per_type_id which holds the ID of the
history entry respective to its type: for "kanban", it is the kanban ID,
for "artifact" it is the artifact ID.

Notes:
- Restler uses @type to understand the type of items held in arrays. It
makes no difference in the API explorer to remove it.
- It is fine to specify the runtime type of properties (not just
docblocks)
- Since we use a private static "factory" method, the constructor just
assigns its parameters to properties. We can do the same in the
method and it's just shorter.

part of story #24971 [modal] search for artifacts

Change-Id: I5e3a24a092565313d670a8857046bdc4e28dbea0