feat: Simplify new version modal for empty
This is a follow-up of commit
git #tuleap/stable/48fe4ee5514b5efca883b2ba024bc9b733230bef, this one
focus on the creation of a new version of an empty document.

The type selector is removed from the modal and the user must select the
right type of item beforehand, via a dropdown.

This dropdown is dispayled as a regular dropdown under a dedicated "+
New version" button in quick look, or as a sub-dropdown when the "+ New
version" is in the existing dropdown (e.g tree view, search).

Part of story #28273: create OnlyOffice documents

Change-Id: I4ca29458018a6dda7dbc793d78cce17ac21e1a95