refactor: Migrate NewInlineCommentForm to Hybrids + TypeScript
part of request #29602 Migrate pullrequest file-diffs to TypeScript

This is a first step in order to extract the most logic out of the
side-by-side-diff-component and to be able to unit test this logic.

The component depending too much on widgets creation, and, the new comment
widget being written in angular.js, it is not possible yet to achieve
that goal without extracting that part out of angular.

[note]

Regarding the workaround getLineNumberFromEvent, it appears that
CodeMirror can provide the wrong line number. We now parse the
textContent of the gutter element we can access from the provided event.

How to test:
No functional change expected. However, please verify the following points:
- When a gutter is clicked (sbs and unified diffs), a
  NewInlineCommentForm is added to the line.
- A placeholder has been created on the opposite line with the right
  height
- When you cancel your comment, the widget is removed, and the opposite
  placeholder is removed (or its height is updated if it was there
  before)
- When you submit, the widget is removed, and a PullRequestComment
  widget is added in place of the previous one.

Change-Id: I06f68d053bb56c52a75f3e147952f4e372e15f93