fix: required indicator in timeframe administration
During git #tuleap/stable/f36c0817b7b4c43e07a59fb6234af8fbc93d35da a bug
has been introduced and has not been spotted by CI, despite a good unit
tests coverage. This is due to the fact that the application is located
deep in the hierarchy, out of reach for `pnpm test`.

The bug is visible for a timeframe based on field: the required
indicator (red asterisk) is not properly displayed for duration field.

Since we want to have independant apps (See
`adr/0017-multiple-apps-per-context.md`), this contribution hits two
birds with one stone (in addition to fix the bug, so probably three
birds 🦆🦆🦆 \o/):

* semantics-timeframe is now an independant App. We will later be able
to convert it to Vue3 without messing with other Vue2 apps in tracker,
* unit tests are now run by the CI (see Patchset 3 that is trying to run
typecheck).

Note: typecheck task is not included in this contribution since it is
out of scope. It is better to add it once we have converted the app to
contribution API IMHO.

[0]: adr/0017-multiple-apps-per-context.md

Part of request #26738: Convert semantics administration to Burning Parrot
Part of request #29264: Split the build of tracker plugin frontend assets into multiple apps

Change-Id: I13feb3c8f403e858f16e7b5d75b9bec80f2d123e