refactor: New inline comment saves itself
part of request #29602 Migrate pullrequest file-diffs to TypeScript

New inline comments are saved through FileDiffRestService.
Since we want to extract the widget creation part from angular, we need
to save the new inline comment differently.

[Note] about error display

We no longer open the error modal when the POST fails. This modal is
opened by an angular service. I'll check with UXD how we are supposed to
display errors when they occurs.

-- no functional change expected

Change-Id: I03105484299335f7bd77fbeafb3cf722587459cc