feat: Change "New version" label for empty documents
"New version" action for an empty document allows a user to convert it
to a "real" document (a file, an embedded, …). To better reflect this
behavior we rename the action to "Convert to…".

"Upload a file" option is renamed to "Uploaded file" to have a better
match: "Convert to… Uploaded file".

Part of story #28273: create OnlyOffice documents

Change-Id: I23124ea70f8f567b94da255b1d330a64fa2d6c85