feat: Build & Run on RockyLinux 9.0
Introduce the development Build & Run. Only a subset of service works
and its mostly to discuss the approach for the management of the build
and the strategy we are going to apply to deal with OS variant.

Still, the generated packages can be install and run with a simple
config.

The SystemD unit service file to serve PHP (for Tuleap & MediaWiki
standalone) were modified to be compliant with el9 version of SystemD:
- EnvironmentFile was pointing to an empty file (useless)
- PIDFile is not recommended for `notify` type of units [1]

[1] https://www.freedesktop.org/software/systemd/man/systemd.service.html#PIDFile=

Part of: request #29619 Add support of Rocky Linux 9

Change-Id: I49f2bc2de06f0eec92caf5f2a72c0593dbc9b841