refactor: use autocompleter even at first load
No functional change expected in the Artifact links field in the modal:
- enable the feature flag

$ tuleap config-set feature_flag_history_and_search_in_modal 1

- after enabling it, when the current link type is NOT "Parent", without
typing anything in the input field, there is an empty "recent items"
section in the dropdown.
- when the current link type is "Parent", possible parents are shown.
- without the feature flag, the dropdown is empty.

No functional change for auto-completion: if you type something, it
filters the list of possible parents. If you type an artifact ID, you
get a matching artifact.

Part of story #24971 [modal] search for artifacts

Change-Id: I29d39095228d73768ec3d930327d2ec72f5a40f8