feat: uniform display of dropdowns
Some dropdowns have `tlp-dropdown-menu-right` classname, some does not.
Every dropdowns in Document should open the same way. We decide to
remove `tlp-dropdown-menu-right` for two reasons:
* It is the default behavior in applications.
* There is less chance to hide list of items (tree or search views) that
user is currently scanning.

We might want to generalize this to the whole Tuleap codebase by
removing `tlp-dropdown-menu-right` from TLP, but it will be part of a
dedicated contribution if needed.

Part of story #28273: create OnlyOffice documents

Change-Id: Ifddae36dd37265fe54c604eab92eae61c763306c